Assignment-2/requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langchain-core==0.2.38
python-dotenv==1.0.1
pydantic==1.10.21
openai==1.40.0